+++ docs/docs/changelog.md
@@ -9,7 +9,8 @@
 - Fixed issue where testing repo wasnt properly detected in dynawidgets
 - Fixed log level functionality -> https://github.com/Panonim/dynacat/pull/96
 - Fixed incorrect daily percentage in market widget -> https://github.com/Panonim/dynacat/pull/117
-- Added server-stats `compact: true` option -> https://github.com/Panonim/dynacat/pull/104 
+- Added server-stats `compact: true` option -> https://github.com/Panonim/dynacat/pull/104
+- Added support for Sonarr/Radarr releases in the `calendar` widget 
 
 # Changes for 2.3.1
 - Added support for loading environment variables from a file via `--env-file`